ml/requirements.txt
pandas
scikit-surprise
SQLAlchemy
psycopg2-binary
Flask